+++ pkg/NEWS.md
@@ -8,6 +8,8 @@
 
 * adapt configure.ac for GEOS 3.10
 
+* add support for buffered geometry fixer in gMakeValid() from 3.10
+
 # Version 0.5-5 (2020-09-07, rev. 634-640)
 
 * add gMakeValid() for GEOS 3.8